+++ src/components/dropdown/dropdown.component.ts
@@ -84,7 +84,7 @@ export class DropdownComponent implements ControlValueAccessor, OnInit {
   private onChangeCallback: (_: any) => void;
 
   writeValue(v: any): void {
-    if (v.toString() != this.value.toStrig())
+    if (v.toString() != this.value.toString())
       this.value = v;
   }
   registerOnChange(fn: any) {