+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## master
 
 ### Breaking Changes
+- vcl-input:
+  - change from directive to component
+  - rename input/outputs to equal html5-input
 - dropdown: see [readme](https://github.com/ng-vcl/ng-vcl/blob/master/src/components/dropdown/README.md)
   - "select"-event is now "change"
   - the select-event returns now an value-array instead of the items
@@ -10,6 +13,10 @@
 - layer:
   - name reference support removed
   - introduce component layers and LayerRefs
+
+### Changes:
+  - added vcl-file-input
+### Bugfixes
 
 ## 15979eb (2016-12-14)